+++ main.py
@@ -1,11 +1,27 @@
 from fastapi import FastAPI
+from fastapi.concurrency import asynccontextmanager
 from fastapi.middleware.cors import CORSMiddleware
+from src.dal.local.redis_adapter import RedisAdapter
 from src.presentation.routes.source_item_route import source_item_router
 from src.presentation.routes.topics_route import app_topics_router
 from src.presentation.routes.input_route import input_router
 from src.presentation.routes.context_route import context_router
 from src.presentation.routes.pdf_route import pdf_router
+from src.core.settings import app_settings
 
+settings = app_settings()
+
+@asynccontextmanager
+async def lifespan(app: FastAPI):
+    # Startup
+    adapter = RedisAdapter(settings.REDIS_URL, namespace=settings.REDIS_NAMESPACE)
+    await adapter.connect()
+    app.state.redis = adapter
+
+    yield  # <---- aqui a app roda normalmente
+
+    # Shutdown
+    await adapter.close()